requirements.txt
openai>=1.50.0
requests>=2.31.0
backoff>=2.2.1
patchright>=1.59.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pandas>=2.0.0

# OCR Pipeline — PaddleOCR (SOTA, table-aware, 111 languages incl Bengali)
paddlepaddle>=3.0.0
paddleocr>=3.5.0
pypdfium2>=5.0.0
pdf2image>=1.17.0
